Split calculator & results into separate components
